// apps/web/src/lib/api-client/models/paginatedTasksResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */
import type { PaginationMetaDto } from './paginationMetaDto';
import type { TaskDto } from './taskDto';

export interface PaginatedTasksResponseDto {
  /** Lista de tarefas paginadas */
  data: TaskDto[];
  /** Metadados de paginação */
  meta: PaginationMetaDto;
}