// apps/web/src/lib/api-client/models/paginatedUsersResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */
import type { ManagedUserDto } from './managedUserDto';
import type { PaginationMetaDto } from './paginationMetaDto';

export interface PaginatedUsersResponseDto {
  data: ManagedUserDto[];
  meta: PaginationMetaDto;
}